Delete old enable review app files & rename new files
